requirements.txt
asyncio
aiohttp
pigpio
requests
paho-mqtt
pychromecast
pyyaml
aiomonitor
netdisco
